requirements.txt
matplotlib==3.9.2
numpy_financial==1.0.0
pandas==2.2.3
Pyrebase4==4.8.0
streamlit==1.24.1
openpyxl
